--- Java/Car.java
@@ -1,7 +1,0 @@
-
-public class Car implements Vehicle {
-  @Override
-  public String getType() {
-    return "Car";
-  }
-}
