+++ Java/src/main/java/com/codetest/model/Car.java
@@ -1,0 +1,8 @@
+package com.codetest.model;
+
+public class Car implements Vehicle {
+  @Override
+  public String getType() {
+    return "Car";
+  }
+}
