+++ Java/src/main/java/com/codetest/model/Motorbike.java
@@ -1,0 +1,8 @@
+package com.codetest.model;
+
+public class Motorbike implements Vehicle {
+  @Override
+  public String getType() {
+    return "Motorbike";
+  }
+}
