+++ Java/src/main/java/com/codetest/model/Vehicle.java
@@ -1,0 +1,6 @@
+package com.codetest.model;
+
+public interface Vehicle {
+
+  public String getType();
+}
